requirements.txt
appnope==0.1.0
DateTime==4.1.1
decorator==4.0.11
ipython==5.3.0
ipython-genutils==0.2.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyserial==3.3
pytz==2017.2
PyYAML==3.12
schedule==0.4.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
zope.interface==4.3.3
